Check for error cases in cmd run func
The cmd main.go file will now be ignored from tests.
The cmd run() method is now in it's own file.
