+++ makefile
@@ -15,10 +15,10 @@ clean:
 	rm -rf bin
 
 test:
-	go test -cover ./...
+	go test -cover -tags test ./...
 
 test-coverage: bin
-	go test -coverprofile=bin/coverage.out ./...
+	go test -coverprofile=bin/coverage.out -tags test ./...
 	go tool cover -html bin/coverage.out
 
 check-env:
